requirements.txt
numpy==1.12.0
Pillow==4.0.0
scipy
tensorflow-gpu==1.0.0
matplotlib
